SDK-408: Extract KeyStreamVisitor for reuse
