+++ yoti-sdk-impl/src/main/java/com/yoti/api/client/spi/remote/SecureYotiClient.java
@@ -1,12 +1,8 @@
 package com.yoti.api.client.spi.remote;
 
-import static com.yoti.api.client.spi.remote.call.YotiConstants.DEFAULT_CHARSET;
 import static com.yoti.api.client.spi.remote.util.Validation.notNull;
 
-import java.io.BufferedReader;
 import java.io.IOException;
-import java.io.InputStream;
-import java.io.InputStreamReader;
 import java.security.KeyPair;
 import java.security.Security;
 
@@ -14,7 +10,6 @@ import com.yoti.api.client.ActivityDetails;
 import com.yoti.api.client.AmlException;
 import com.yoti.api.client.InitialisationException;
 import com.yoti.api.client.KeyPairSource;
-import com.yoti.api.client.KeyPairSource.StreamVisitor;
 import com.yoti.api.client.ProfileException;
 import com.yoti.api.client.YotiClient;
 import com.yoti.api.client.aml.AmlProfile;
@@ -22,9 +17,6 @@ import com.yoti.api.client.aml.AmlResult;
 import com.yoti.api.client.spi.remote.call.Receipt;
 import com.yoti.api.client.spi.remote.call.aml.RemoteAmlService;
 
-import org.bouncycastle.openssl.PEMKeyPair;
-import org.bouncycastle.openssl.PEMParser;
-import org.bouncycastle.openssl.jcajce.JcaPEMKeyConverter;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -76,29 +68,6 @@ final class SecureYotiClient implements YotiClient {
         } catch (IOException e) {
             throw new InitialisationException("Cannot load key pair", e);
         }
-    }
-
-    private static class KeyStreamVisitor implements StreamVisitor {
-
-        @Override
-        public KeyPair accept(InputStream stream) throws IOException, InitialisationException {
-            PEMParser reader = new PEMParser(new BufferedReader(new InputStreamReader(stream, DEFAULT_CHARSET)));
-            KeyPair keyPair = findKeyPair(reader);
-            if (keyPair == null) {
-                throw new InitialisationException("No key pair found in the provided source");
-            }
-            return keyPair;
-        }
-
-        private KeyPair findKeyPair(PEMParser reader) throws IOException {
-            for (Object o; (o = reader.readObject()) != null; ) {
-                if (o instanceof PEMKeyPair) {
-                    return new JcaPEMKeyConverter().getKeyPair((PEMKeyPair) o);
-                }
-            }
-            return null;
-        }
-
     }
 
 }
